requirements.txt
einops
flax
datasets
transformers
torchvision
pytorch-lightning
omegaconf
wandb
